pyproject.toml
[tool.poetry]
name = "pygmoss"
version = "0.1.0"
description = "Global Model for the Radio Sky Spectrum"
authors = ["Dr Mayuri Rao, Mohith P. A. <mohithpa2@gmail.com>"]
readme = "README.md"
packages = [{include = "pygmoss"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
scipy = "^1.11.3"
numpy = "^1.26.0"
matplotlib = "^3.8.0"
healpy = "^1.16.6"
pandas = "^2.1.1"
pytest = "^7.4.2"
poetry = "^1.6.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
